requirements.txt
facenet_pytorch==2.5.0
joblib==0.17.0
librosa==0.8.0
matplotlib==3.3.2
mmcv==0.4.3
numpy==1.19.2
opencv_python==4.4.0.44
pandas==1.1.3
Pillow==8.4.0
pytorch_lightning==1.3.3
scikit_learn==1.0.1
torch==1.9.0
torchmetrics==0.2.0
torchvision==0.10.0
wandb==0.10.30
